+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 ## 3.1.0 - 2021-03-12
+### Added
+- Linux now has a packaged version, and appropriate icons on the update screen
+
 ### Fixed
 - Fixed race condition when opening update window on Linux